+++ src/trx/game/objects/creatures/bat.c
@@ -31,7 +31,7 @@ static void M_FixEmbeddedPosition(int16_t item_num)
 
     int16_t room_num = item->room_num;
     const SECTOR *const sector = Room_GetSector(item->pos, &room_num);
-    const int16_t ceiling = Room_GetCeiling(sector, item->pos);
+    const int32_t ceiling = Room_GetCeiling(sector, item->pos);
 
     // The bats animation and frame have to be changed to the hanging
     // one to properly measure them. Save it so it can be restored
